requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0.0
requests>=2.31.0
django-allauth[socialaccount]>=6.0.0
